# WebService/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A Mini Product Ordering Service API.",
        "title": "Gin Mini Product Ordering Service",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/Product/AddCategory": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Add Category Product",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Product Management"
                ],
                "summary": "Add Category Product",
                "parameters": [
                    {
                        "description": "Add Category Product",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ProductArr"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/Product/AddImage": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Product Image Form",
                "consumes": [
                    "multipart/form-data"
                ],
                "tags": [
                    "Product Management"
                ],
                "summary": "Product Image Form",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "image",
                        "name": "image",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/Product/AddProduct": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Add Product",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Product Management"
                ],
                "summary": "Add Product",
                "parameters": [
                    {
                        "description": "Add Product ",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Product"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/Product/AddStock": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Add Stock Product",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Product Management"
                ],
                "summary": "Add Stock Product",
                "parameters": [
                    {
                        "description": "Add Stock Product",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ProductStock"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/Product/Delete/{pid}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete Product",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Product Management"
                ],
                "summary": "Delete Product",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Delete Product by ID",
                        "name": "pid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/Product/RemoveCategory": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Remove Category Product",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Product Management"
                ],
                "summary": "Remove Category Product",
                "parameters": [
                    {
                        "description": "Remove Category Product",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ProductArr"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
        "/Product/UpdateProduct": {
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Add Product",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Product Management"
                ],
                "summary": "Add Product",
                "parameters": [
                    {
                        "description": "Update Product",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateProduct"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
    },
    "definitions": {
...
        "models.Product": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "buku",
                        "alat tulis",
                        "keren",
                        "sekolah"
                    ]
                },
                "desc": {
                    "type": "string",
                    "example": "Ini Buku Ajaib"
                },
                "name": {
                    "type": "string",
                    "example": "Buku"
                },
                "price": {
                    "type": "integer",
                    "example": 2000
                },
                "stock": {
                    "type": "integer",
                    "example": 5000
                }
            }
        },
        "models.ProductArr": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "buku",
                        "alat tulis",
                        "keren",
                        "sekolah"
                    ]
                },
                "id": {
                    "type": "string"
                }
            }
        },
...
        "models.ProductStock": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "stock": {
                    "type": "integer"
                }
            }
        },
...
        "models.UpdateProduct": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "buku",
                        "alat tulis",
                        "keren",
                        "sekolah"
                    ]
                },
                "desc": {
                    "type": "string",
                    "example": "Ini Buku Ajaib"
                },
                "id": {
                    "type": "string",
                    "example": "9232131"
                },
                "image": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string",
                    "example": "Buku"
                },
                "price": {
                    "type": "integer",
                    "example": 2000
                },
                "stock": {
                    "type": "integer",
                    "example": 5000
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Token",
            "in": "header"
        }
    }
}